+++ rv/modules/sampler.py
@@ -290,6 +290,7 @@ class Sampler(Module):
     record_with_reduced_sample_rate = Option(False)
     record_in_16_bit = Option(False)
     stop_recording_on_project_stop = Option(False)
+    ignore_velocity_for_volume = Option(False)
 
     def __init__(self, **kwargs):
         super(Sampler, self).__init__(**kwargs)