+++ rv/__init__.py
@@ -1,5 +1,5 @@
 """Radiant Voices"""
 
-__version__ = '0.4.0'
+__version__ = '0.4.0.dev1'
 
 ENCODING = 'utf8'